requirements.txt
numpy==1.25.0
pandas==1.5.3
Pillow==9.5.0
streamlit==1.28.1
